WIP: Location feed

+++ Forecasting/Model/GpsCoords.swift
@@ -7,10 +7,10 @@
 //
 
 struct GpsCoords {
-    let latitude: Float
-    let longitude: Float
+    let latitude: Double
+    let longitude: Double
     
-    init(latitude: Float, longitude: Float) {
+    init(latitude: Double, longitude: Double) {
         self.latitude = latitude
         self.longitude = longitude
     }
